requirements.txt
tensorflow==2.12.0
numpy==1.23.5
pandas==2.0.2
matplotlib==3.7.1
click==8.1.7
Pillow==9.5.0
scikit-image==0.22.0
keras==2.12.0